% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_4to5_week_regressors.r
\name{gen_4to5_week_regressors}
\alias{gen_4to5_week_regressors}
\title{generate 4/5 week regressors}
\usage{
gen_4to5_week_regressors(
  start_year = NULL,
  end_year = NULL,
  omit_march = TRUE,
  remove_cal_means = FALSE
)
}
\arguments{
\item{start_year}{First year of the sequence
This is a required argument.}

\item{end_year}{Ending year of the sequence
This is a required argument.}

\item{omit_march}{Logical scalar, exclude March from the set of regressors returned
Default is TRUE}

\item{remove_cal_means}{Logical scalar;
if TRUE, calendar month means are removed from the final regression matrix.
Default setting is FALSE.}
}
\value{
Matrix of time series arrays of 4/5 week regressors starting in January of
\code{start_year} and ending in December of \code{end_year}, with each column
regresenting a different month.
}
\description{
Generate 4 to 5 week effect regressors from paper by Cano, Scott, Kropf, Scott and Stamas (1996)
}
\details{
Version 3.0 5/6/2024
}
\examples{
four2five_reg     <- gen_4to5_week_regressors(2005, 2024)
four2five_reg_cal <- gen_4to5_week_regressors(2005, 2024, remove_cal_means = TRUE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
