% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_indirect_quarterly_holiday.r
\name{gen_indirect_quarterly_holiday}
\alias{gen_indirect_quarterly_holiday}
\title{Generate indirect quarterly holiday adjustments}
\usage{
gen_indirect_quarterly_holiday(this_a1 = NULL, this_hol = NULL, ratio = TRUE)
}
\arguments{
\item{this_a1}{Real array; ts object of the original series
This is a required argument.}

\item{this_hol}{Real array; ts object of the holiday factors
This is a required argument.}

\item{ratio}{Logical scalar; if TRUE, holiday factors are assumed to be ratios; otherwise, the factors are assumed
to be on the same scale as the original series. Default setting is TRUE.}
}
\value{
List object of two ts objects: \code{holadj}, which contains the indirect holiday adjusted quarterly series and
\code{holfac}, the indirect holiday factors.
}
\description{
Generate indirect quarterly holiday factors and an indirect holiday adjusted series from monthly time series and
monthly holiday factors.
}
\details{
Version 2.0, 5/6/2024
}
\examples{
n2033157_hol_q_list <- gen_indirect_quarterly_holiday(n2033157_a1, n2033157_hol)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
