% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reference_day.r
\name{gen_reference_day}
\alias{gen_reference_day}
\title{Reference Day Generation}
\usage{
gen_reference_day(
  start_year = NULL,
  end_year = NULL,
  census_adj = TRUE,
  back_dates = NULL
)
}
\arguments{
\item{start_year}{First year of the sequence
This is a required argument.}

\item{end_year}{Ending year of the sequence
This is a required argument.}

\item{census_adj}{Logical scalar, apply adjustment done by US Census Bureau for November and December.
Default is TRUE.}

\item{back_dates}{Array of date objects where the reference week needs to be set back a week by interviewers}
}
\value{
Array of reference week dates starting in January of \code{start_year} and ending in December of \code{end_year}.
}
\description{
Generate the start of the reference week for a span of years, with an option to incorporate exceptions
}
\details{
Version 3/0 5/6/2024
}
\examples{
replacement_dates <- c(as.Date('2013-11-10'), as.Date('2019-11-10'))
rw2005 <- gen_reference_day(2005, 2024, back_dates = replacement_dates)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
