% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_reference_week_start_regressor.r
\name{gen_reference_week_start_regressor}
\alias{gen_reference_week_start_regressor}
\title{generate reference week start holiday regressors}
\usage{
gen_reference_week_start_regressor(
  this_reference_day = NULL,
  collapse_col = NULL,
  reg_means = NULL,
  contrast_reg = TRUE
)
}
\arguments{
\item{this_reference_day}{Array of date objects of the start of the reference week for a given month
This is a required argument.}

\item{collapse_col}{integer scalar; collapses the first \code{collapse_col} columns into a single column;
Default setting is NULL, all columns are returned.}

\item{reg_means}{numeric vector; vector of means to be removed from the regressors;
Default is NULL, no mean removal done.}

\item{contrast_reg}{Logical scalar; if TRUE, contrast regressors are generated.
Default setting is TRUE.}
}
\value{
Time series regression matrix object with reference week start regressors.
}
\description{
Generate reference week start holiday regressors, defined as a time series matrix.
}
\details{
Version 4.0 5/6/2024
}
\examples{
replacement_dates <- c(as.Date('2013-11-10'), as.Date('2019-11-10'))
rw2005 <- gen_reference_day(2005, 2024, back_dates = replacement_dates)
ref_week_start_reg <- gen_reference_week_start_regressor(rw2005)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
