% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rw_holiday_matrix.r
\name{gen_rw_holiday_matrix}
\alias{gen_rw_holiday_matrix}
\title{Reference week regression matrix}
\usage{
gen_rw_holiday_matrix(
  this_reference_week = NULL,
  add_gf = TRUE,
  add_easter = TRUE,
  add_labor = TRUE,
  add_columbus = TRUE,
  add_vet = TRUE,
  remove_cal_means = TRUE
)
}
\arguments{
\item{this_reference_week}{Array of date objects of the start of the reference week for a given month
This is a required argument.}

\item{add_gf}{Logical scalar; if TRUE, a Good Friday holiday regressor will be included in the regression matrix.
Default setting is TRUE.}

\item{add_easter}{Logical scalar; if TRUE, an Easter holiday regressor will be included in the regression matrix.
Default setting is TRUE.}

\item{add_labor}{Logical scalar; if TRUE, a Labor Day holiday regressor will be included in the regression matrix.
Default setting is TRUE.}

\item{add_columbus}{Logical scalar; if TRUE, a Columbus Day holiday regressor will be included in the regression matrix.
Default setting is TRUE.}

\item{add_vet}{Logical scalar; if TRUE, a Veteran's Day holiday regressor will be included in the regression matrix.
Default setting is TRUE.}

\item{remove_cal_means}{Logical scalar; if TRUE, calendar month means are removed from the final regression matrix.
Default setting is TRUE.}
}
\value{
Array of reference week dates starting in January of \code{start_year} and ending in December of \code{end_year}
}
\description{
Generate full regression matrix for reference week related holiday regressors, defined as a time series object.
}
\details{
Version 2.0, 5/6/2024
}
\examples{
replacement_dates <- c(as.Date('2013-11-10'), as.Date('2019-11-10'))
rw2005 <- gen_reference_day(2005, 2024, back_dates = replacement_dates)
holiday_matrix_2005 <- gen_rw_holiday_matrix(rw2005, add_gf = FALSE, add_easter = FALSE,
                       remove_cal_means = FALSE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
