% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rw_regressors.r
\name{gen_rw_regressors}
\alias{gen_rw_regressors}
\title{Generate reference week regressor}
\usage{
gen_rw_regressors(
  holiday_years = NULL,
  holiday_month = NULL,
  start_year = NULL,
  end_year = NULL,
  col_label = "Reg",
  join_regressors = TRUE,
  remove_cal_means = TRUE
)
}
\arguments{
\item{holiday_years}{Integer array; Years where the holiday appears in the reference week (or for Columbus Day
or Veteran's day, years where the holiday does not occur in the reference week)
This is a required argument.}

\item{holiday_month}{Integer scalar; The month in which this holiday occurs. For Good Friday and Easter,
this would be April (4); for Labor Day, this would be September (9), etc.
This is a required argument.}

\item{start_year}{Integer scalar; First year of the generated holiday regressor.
The regressor will begin on the first observation of this year.
This is a required argument.}

\item{end_year}{Integer scalar; Final year of the generated holiday regressor.
The regressor will end on the last observation of this year.
This is a required argument.}

\item{col_label}{Character scalar; A label used to generate column names for the individual AO regressors
if \code{join_regressors = FALSE}.}

\item{join_regressors}{Logical scalar;
if TRUE, individual indicator regressors are combined into one grouped regressor.
If FALSE, a matrix of the individual AO regressors will be returned.
Default setting is TRUE.}

\item{remove_cal_means}{Logical scalar; if TRUE, calendar month means are removed from the
final regression matrix.
Default setting is TRUE.}
}
\value{
if \code{join_regressors = TRUE}, a time series object with the holiday regressor is returned;
otherwise, a matrix of AO regressors for the individual holidays is returned.
}
\description{
Generate specific reference week related holiday regressors, defined as a time series object.
The object returned is either a time series vector or a matrix of time series indicator variables.
}
\details{
Version 3.0, 5/6/2024
}
\examples{
gf_years <- c(2001, 2006, 2017, 2022)
# returns a time series object with one grouped regressor
this_gf_grouped <-
     gen_rw_regressors(gf_years, 4, 2000, 2027)
# returns a time series matrix with four columns of indicator regressors
this_gf_individual <-
     gen_rw_regressors(gf_years, 4, 2000, 2027, 'gf', join_regressors = FALSE)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
