% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2033157_a1.r
\docType{data}
\name{n2033157_a1}
\alias{n2033157_a1}
\title{At Work Series}
\format{
A time series object of an at work hours series from January of 2003 to May of 2020
}
\usage{
n2033157_a1
}
\description{
A time series object of an at work hours series
}
\keyword{datasets}
