% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2033157_hol.r
\docType{data}
\name{n2033157_hol}
\alias{n2033157_hol}
\title{At Work Series Monthly Holiday Factors}
\format{
A time series object of monthly holiday factors from an at work hours series from January of 2003 to May of 2020
}
\usage{
n2033157_hol
}
\description{
A time series object of monthly holiday factors from an at work hours series
}
\keyword{datasets}
