% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_user_reg.r
\name{save_user_reg}
\alias{save_user_reg}
\title{save user regression matrix}
\usage{
save_user_reg(this_reg = NULL, this_reg_file = NULL)
}
\arguments{
\item{this_reg}{double precision time series array or matrix to be saved.
This is a required argument.}

\item{this_reg_file}{character string; name of file time series array or matrix to be saved to.
This is a required argument.}
}
\value{
file with user-defined regressors will be produced.
}
\description{
Save a user-defined regression array or matrix with time series attributes to an external ASCII file
in X-13ARIMA-SEATS' datevalue format
}
\details{
Version 2.0, 5/6/2024
}
\examples{
gf_years <- c(2001, 2006, 2017, 2022)
this_gf <-
     gen_rw_regressors(gf_years, 4, 2000, 2027, 'gf', remove_cal_means = FALSE)
\dontrun{save_user_reg(this_gf, 'gf_2000_2027.txt')}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
